Add temporary migration scripts for the folder structure
These will be removed, before we merge the branch.
We can test the new folder structure from this branch easily now,
without checking in the changed folder structure (that would create
a mess of conflicts and should only be done at the very end when
everything is tested and we want to roll it out to master).

Right now, this contains my proposed folder structure from #194
and this is subject to change (please write into that ticket if you
prefer another folder structure).